requirements.txt
discord.py
python-dotenv
tts==0.10.2